+++ alcs-frontend/src/app/shared/utils/boolean-helper.ts
@@ -1,3 +1,16 @@
+export const parseStringToBoolean = (val?: string | null) => {
+  switch (val) {
+    case 'true':
+      return true;
+    case 'false':
+      return false;
+    case null:
+      return null;
+    default:
+      return undefined;
+  }
+};
+
 export const parseBooleanToString = (val?: boolean | null) => {
   switch (val) {
     case true:
