--- alcs-frontend/src/app/shared/utils/string-helper.ts
@@ -1,12 +1,0 @@
-export const parseStringToBoolean = (val?: string | null) => {
-  switch (val) {
-    case 'true':
-      return true;
-    case 'false':
-      return false;
-    case null:
-      return null;
-    default:
-      return undefined;
-  }
-};
